udao/requirements.txt
mypy==1.5.1
mypy-extensions==1.0.0
networkx==2.8.8
networkx-stubs==0.0.1
numpy==1.26.1
pandas==2.1.1
pandas-stubs==2.1.1.230928
pre-commit==3.4.0
pytest==7.4.2
pytest-cov==4.1.0
pytest==7.4.2
pytest-mock==3.12.0
scikit-learn==1.3.1
pytorch-warmup==0.1.1
lightning==2.1.0
torchmetrics==1.2.0
gensim==4.3.0
matplotlib==3.8.1
Platypus-Opt==1.1.0
